Swoosh tested across clang, GNU, and MSVC. Fixes made.

+++ src/Segues/PixelateBlackWashFade.h
@@ -1,7 +1,7 @@
 #pragma once
-#include <Swoosh\Segue.h>
-#include <Swoosh\Ease.h>
-#include <Swoosh\EmbedGLSL.h>
+#include <Swoosh/Segue.h>
+#include <Swoosh/Ease.h>
+#include <Swoosh/EmbedGLSL.h>
 
 using namespace swoosh;
 
@@ -56,7 +56,7 @@ public:
 
       sf::RectangleShape whiteout;
       whiteout.setSize(sf::Vector2f((float)surface.getTexture().getSize().x, (float)surface.getTexture().getSize().y));
-      whiteout.setFillColor(sf::Color(0, 0, 0, (sf::Uint8)(alpha * 255.0)));
+      whiteout.setFillColor(sf::Color(0, 0, 0, (sf::Uint8)(alpha * (double)255)));
       surface.draw(whiteout);
     }
   }
